moved out request unittests
