+++ automower_ble/request.py
@@ -290,209 +290,3 @@ class MowerRequest:
 
         return self.__finalise_standard_request(data)
 
-
-class TestStringMethods(unittest.TestCase):
-    def test_generate_request_setup_channel_id(self):
-        request_one = MowerRequest(1197489075)
-        request_two = MowerRequest(1739453030)
-
-        self.assertEqual(
-            request_one.generate_request_setup_channel_id(),
-            bytearray(
-                [
-                    0x02,
-                    0xFD,
-                    0x16,
-                    0x00,
-                    0x00,
-                    0x00,
-                    0x00,
-                    0x00,
-                    0x00,
-                    0x2E,
-                    0x14,
-                    0xB3,
-                    0x3B,
-                    0x60,
-                    0x47,
-                    0x00,
-                    0x00,
-                    0x00,
-                    0x00,
-                    0x4D,
-                    0x61,
-                    0x69,
-                    0x6E,
-                    0x00,
-                    0x1B,
-                    0x03,
-                ]
-            ),
-        )
-        self.assertEqual(
-            binascii.hexlify(request_one.generate_request_handshake()),
-            b"02fd0a00b33b6047005d08012803",
-        )
-
-        self.assertEqual(
-            binascii.hexlify(request_two.generate_request_setup_channel_id()),
-            b"02fd160000000000002e1466f2ad67000000004d61696e003403",
-        )
-        self.assertEqual(
-            binascii.hexlify(request_two.generate_request_handshake()),
-            b"02fd0a0066f2ad6700f908012803",
-        )
-
-    def test_generate_request_device_type(self):
-        request_two = MowerRequest(1739453030)
-
-        self.assertEqual(
-            binascii.hexlify(request_two.generate_request_device_type()),
-            b"02fd100066f2ad6701d700af5a1209000000b703",
-        )
-
-    def test_generate_request_pin(self):
-        request_two = MowerRequest(1739453030)
-
-        self.assertEqual(
-            request_two.generate_request_pin(7201),
-            bytearray(
-                [
-                    0x02,
-                    0xFD,
-                    0x12,
-                    0x00,
-                    0x66,
-                    0xF2,
-                    0xAD,
-                    0x67,
-                    0x01,
-                    0xAD,
-                    0x00,
-                    0xAF,
-                    0x38,
-                    0x12,
-                    0x04,
-                    0x00,
-                    0x02,
-                    0x00,
-                    0x21,
-                    0x1C,
-                    0x7D,
-                    0x03,
-                ]
-            ),
-        )
-
-    def test_generate_request_remaining_charge_time(self):
-        request_one = MowerRequest(1197489075)
-
-        self.assertEqual(
-            request_one.generate_request_remaining_charge_time(),
-            bytearray(
-                [
-                    0x02,
-                    0xFD,
-                    0x10,
-                    0x00,
-                    0xB3,
-                    0x3B,
-                    0x60,
-                    0x47,
-                    0x01,
-                    0x73,
-                    0x00,
-                    0xAF,
-                    0x0A,
-                    0x10,
-                    0x16,
-                    0x00,
-                    0x00,
-                    0x00,
-                    0xB8,
-                    0x03,
-                ]
-            ),
-        )
-
-    def test_generate_request_is_charging(self):
-        request_one = MowerRequest(1197489075)
-
-        self.assertEqual(
-            request_one.generate_request_is_charging(),
-            bytearray(
-                [
-                    0x02,
-                    0xFD,
-                    0x10,
-                    0x00,
-                    0xB3,
-                    0x3B,
-                    0x60,
-                    0x47,
-                    0x01,
-                    0x73,
-                    0x00,
-                    0xAF,
-                    0x0A,
-                    0x10,
-                    0x15,
-                    0x00,
-                    0x00,
-                    0x00,
-                    0x30,
-                    0x03,
-                ]
-            ),
-        )
-
-    def test_generate_request_battery_level(self):
-        request_one = MowerRequest(1197489075)
-
-        self.assertEqual(
-            request_one.generate_request_battery_level(),
-            bytearray(
-                [
-                    0x02,
-                    0xFD,
-                    0x10,
-                    0x00,
-                    0xB3,
-                    0x3B,
-                    0x60,
-                    0x47,
-                    0x01,
-                    0x73,
-                    0x00,
-                    0xAF,
-                    0x0A,
-                    0x10,
-                    0x14,
-                    0x00,
-                    0x00,
-                    0x00,
-                    0xBF,
-                    0x03,
-                ]
-            ),
-        )
-
-    def test_generate_request_mode_of_operation(self):
-        request = MowerRequest(0x5798CA1A)
-
-        self.assertEqual(
-            binascii.hexlify(request.generate_request_mode_of_operation("manual")),
-            b"02fd11001aca9857013400afea1100000100003303",
-        )
-
-    def test_generate_request_override_duration(self):
-        request = MowerRequest(0x5798CA1A)
-
-        self.assertEqual(
-            binascii.hexlify(request.generate_request_override_duration("3hours")),
-            b"02fd14001aca985701fd00af321203000400302a00004603",
-        )
-
-
-if __name__ == "__main__":
-    unittest.main()
